# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LFG Energy Mapping API - OpenStreetMap energy infrastructure viewer and editor
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type EnergyFeatureOsmType = typeof EnergyFeatureOsmType[keyof typeof EnergyFeatureOsmType];


export const EnergyFeatureOsmType = {
  node: 'node',
  way: 'way',
  relation: 'relation',
} as const;

export interface GeoGeometry {
  type?: string;
  coordinates?: unknown[];
}

export interface OsmTags {[key: string]: string}

export interface EnergyFeature {
  /** Unique identifier (osmType + osmId) */
  id: string;
  osmType: EnergyFeatureOsmType;
  osmId: string;
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lon?: number | null;
  geometry?: GeoGeometry;
  tags: OsmTags;
  /** The value of the power= tag (line, substation, generator, tower, plant, cable) */
  powerType?: string;
  /**
     * Generator source (solar, wind, nuclear, gas, coal, oil, hydro, etc.)
     * @nullable
     */
  generatorSource?: string | null;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  version?: number | null;
  /** @nullable */
  lastEdited?: string | null;
}

export interface TypeCount {
  type: string;
  count: number;
  /** @nullable */
  label?: string | null;
}

export interface EnergyStats {
  totalFeatures: number;
  byType: TypeCount[];
  byGeneratorSource?: TypeCount[];
}

/**
 * @nullable
 */
export type RecentEditBbox = { [key: string]: unknown } | null;

export interface RecentEdit {
  changesetId: string;
  userId: string;
  username: string;
  createdAt: string;
  comment: string;
  /** @nullable */
  featuresEdited?: number | null;
  /** @nullable */
  bbox?: RecentEditBbox;
}

export interface OsmUser {
  id: string;
  username: string;
  displayName?: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  changesetCount?: number | null;
}

export interface Changeset {
  id: string;
  comment?: string;
  createdAt?: string;
}

export interface ChangesetInput {
  comment: string;
  source?: string;
}

export interface TagUpdate {
  changesetId: string;
  tags: OsmTags;
}

export type GetEnergyFeaturesParams = {
south: number;
west: number;
north: number;
east: number;
/**
 * Comma-separated list of power types to filter (line,substation,generator,tower,plant,cable)
 */
types?: string;
};

export type GetRecentEditsParams = {
limit?: number;
};

export type OsmLoginCallbackParams = {
code: string;
state: string;
};