# lib/api-zod/src/generated/types/changeset.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LFG Energy Mapping API - OpenStreetMap energy infrastructure viewer and editor
 * OpenAPI spec version: 0.1.0
 */

export interface Changeset {
  id: string;
  comment?: string;
  createdAt?: Date;
}