# lib/api-zod/src/generated/types/energyFeature.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LFG Energy Mapping API - OpenStreetMap energy infrastructure viewer and editor
 * OpenAPI spec version: 0.1.0
 */
import type { EnergyFeatureOsmType } from './energyFeatureOsmType';
import type { GeoGeometry } from './geoGeometry';
import type { OsmTags } from './osmTags';

export interface EnergyFeature {
  /** Unique identifier (osmType + osmId) */
  id: string;
  osmType: EnergyFeatureOsmType;
  osmId: string;
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lon?: number | null;
  geometry?: GeoGeometry;
  tags: OsmTags;
  /** The value of the power= tag (line, substation, generator, tower, plant, cable) */
  powerType?: string;
  /**
     * Generator source (solar, wind, nuclear, gas, coal, oil, hydro, etc.)
     * @nullable
     */
  generatorSource?: string | null;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  version?: number | null;
  /** @nullable */
  lastEdited?: Date | null;
}