# lib/api-zod/src/generated/types/energyStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LFG Energy Mapping API - OpenStreetMap energy infrastructure viewer and editor
 * OpenAPI spec version: 0.1.0
 */
import type { TypeCount } from './typeCount';

export interface EnergyStats {
  totalFeatures: number;
  byType: TypeCount[];
  byGeneratorSource?: TypeCount[];
}